% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vectors.R
\name{check_len}
\alias{check_len}
\title{Check vectors' lenghts}
\usage{
check_len(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
if x and y have the same length return TRUE. Otherwise the
output will be FALSE
}
\description{
\code{check_len} checks that x and y have the same length
}
\examples{

# same length
x <- 1:3
y <- 2:4
check_len(x, y)

# different length
x <- 1:4
y <- 1:2
check_len(x, y)
}
