% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vectors.R
\name{check_num}
\alias{check_num}
\title{Check regression vectors}
\usage{
check_num(x)
}
\arguments{
\item{x}{an input}
}
\value{
if \code{x} is a numeric vector without NA values return \code{TRUE}.
Otherwise the output will be \code{FALSE}.
}
\description{
\code{check_num} checks that \code{x} is a numeric vector without
\code{NA} values.
}
\examples{
# x is NULL
check_num(NULL)

# x is a list
check_num(list(1, 2, 3))

# x is a dataframe
check_num(data.frame(x=c(1, 2, 3)))

# x is a numeric vector with NA
check_num(c(1, 2, NA))

# x is a  numeric vector
check_num(1:4)
}
