% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vectors.R
\name{check_vectors}
\alias{check_vectors}
\title{Check observed and predicted vectors}
\usage{
check_vectors(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\value{
TRUE if the two vectors are valid, otherwise false.
}
\description{
\code{check_vectors} checks if observed and predicted vectors are numeric
without NA values and that have the same length.
}
\examples{
 y_obs  <- c(3.0, -0.5, 2.0, 5.0)
 y_pred <- c(3.9, -0.2, 2.8, 4.5)
 check_vectors(y_obs, y_pred)
}
