% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{error_metric}
\alias{error_metric}
\title{Compute regression error metric}
\usage{
error_metric(observed, predicted, metric = c("RMSE", "MAE", "R2", "MBE"))
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}

\item{metric}{one of 'rmse', 'mae', 'r2', 'mbe'.}
}
\value{
A numeric vector of length one. A number if
\code{observed} and  \code{predicted} are numeric vectors without NA values
with the same length. Otherwise the output will be \code{NA}.
}
\description{
\code{error_metric} returns a selected error metric between values predicted
by a model and the values actually observed.
}
\details{
The available error metrics are:
\itemize{
  \item Root-mean-square error, \code{rmse}.
  \item Mean absolute error, \code{mae}.
  \item Mean bias error, \code{mbe}.
  \item Coefficient of determination, \code{r2}.
}
}
\examples{

y_obs  <- c(0.22,  0.83, -0.12, 0.89, -0.23, -1.30, -0.15, -1.4,
            0.62,  0.99, -0.18, 0.32,  0.34, -0.30,  0.04, -0.87,
            0.55, -1.30, -1.15, 0.20)
y_pred <- c(0.24,  0.78, -0.66,  0.53, 0.70, -0.75, -0.41, -0.43,
            0.49,  0.79, -1.19,  0.06, 0.75, -0.07,  0.43, -0.42,
           -0.25, -0.64, -1.26, -0.07)
# RMSE
error_metric(y_obs, y_pred, metric = 'rmse')

# MAE
error_metric(y_obs, y_pred, metric = 'mae')

# R2
error_metric(y_obs, y_pred, metric = 'R2')

# MBE
error_metric(y_obs, y_pred, metric = 'mbe')
}
