% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{exp_var_score}
\alias{exp_var_score}
\title{Explained variance score}
\usage{
exp_var_score(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
It provides a measure of how well future samples are likely to be predicted
by the model. The Explained variance best possible score is 1.0, lower
values are worse.
}
