% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{expl_var}
\alias{expl_var}
\title{Explained variance score}
\usage{
expl_var(observed, predicted)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}
}
\value{
A numeric vector of length one. A number in \eqn{(-\infty ,1]} if
\code{observed} and \code{predicted} are numeric vectors without NA values,
with the same length. Otherwise the output will be \code{NA}.

\deqn{EVS = 1 - \frac{var(\hat {y_i} - y_i)}  {var(y_i)}}
where \eqn{\hat {y_i}} is the predicted value of \eqn{y_i}.
}
\description{
\code{expl_var} returns a measure of how well future samples are likely to be
predicted by the model. The Explained variance best possible score is 1.0,
lower values are worse.
}
\examples{
y_obs  <- c(3.0, -0.5, 2.0, 5.0)
y_pred <- c(3.9, -0.2, 2.8, 4.5)
expl_var(y_obs, y_pred)
}
