% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
Mean absolute error is a risk metric
corresponding to the expected value of the absolute error loss or l1-norm
loss.
}
