% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{mbe}
\alias{mbe}
\title{Mean bias error}
\usage{
mbe(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector

It describes the direction of the error bias. Its value, however, is related
to magnitude of values under investigation. A negative MBE occurs when
predictions are smaller in value than observations.}
}
\description{
Mean bias error
}
