% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{msle}
\alias{msle}
\title{Mean squared logarithmic error}
\usage{
msle(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
It provides a risk metric corresponding to
the expected value of the squared logarithmic (quadratic) error or loss.
}
