% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{nmse}
\alias{nmse}
\title{Normalized mean squared error}
\usage{
nmse(observed, predicted)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}
}
\value{
A numeric vector of length one. A number in \eqn{[0, +\infty)} if
\code{observed} and  \code{predicted} are numeric vectors without NA values
with the same length. Otherwise the output will be \code{NA}.
}
\description{
\code{nmse} returns the normalized mean squared error
between values predicted by a model and the values actually observed.
}
\details{
\deqn{ NMSE = \frac{MSE(y_{i}, \hat {y_i})}{MSE(y_{i}, 0)}}
where  \eqn{\hat {y_i}} is the predicted value of \eqn{y_i}.
}
\examples{
 y_obs  <- c(3.0, -0.5, 2.0, 5.0)
 y_pred <- c(3.9, -0.2, 2.8, 4.5)
 nmse(y_obs, y_pred)
}
