% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{nse}
\alias{nse}
\title{Normalized squared error}
\usage{
nse(observed, predicted)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}
}
\value{
A numeric vector of length one. A number in \eqn{[0, +\infty)} if
\code{observed} and  \code{predicted} are numeric vectors without NA values
with the same length. Otherwise the output will be \code{NA}.
}
\description{
\code{nse} returns the normalized mean squared error
between values predicted by a model and the values actually observed.
}
\details{
\deqn{ NSE = \frac{MSE(y_{i}, \hat {y_i})}{MSE(y_{i}, 0)}}
where  \eqn{\hat {y_i}} is the predicted value of \eqn{y_i}.
}
\examples{
 y_obs  <- c(0.22,  0.83, -0.12, 0.89, -0.23, -1.30, -0.15, -1.4,
             0.62,  0.99, -0.18, 0.32,  0.34, -0.30,  0.04, -0.87,
             0.55, -1.30, -1.15, 0.20)
 y_pred <- c(0.24,  0.78, -0.66,  0.53, 0.70, -0.75, -0.41, -0.43,
               0.49,  0.79, -1.19,  0.06, 0.75, -0.07,  0.43, -0.42,
              -0.25, -0.64, -1.26, -0.07)
 nse(y_obs, y_pred)
}
