% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg.R
\name{plotPO}
\alias{plotPO}
\title{Create a plot of observed versus predicted values}
\usage{
plotPO(observed, predicted, add_smooth = FALSE)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}

\item{add_smooth}{add smooth line, boolena.}
}
\value{
a ggplot class
}
\description{
Create a plot of observed versus predicted values
}
\examples{
 y_obs  <- c(3.0, -0.5, 2.0, 5.0)
 y_pred <- c(3.1, -0.4, 1.8, 4.5)
 plotPO(y_obs, y_pred)
}
