% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{r2}
\alias{r2}
\title{Coefficient of determination}
\usage{
r2(observed, predicted)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}
}
\value{
A numeric vector of length one. A number in \eqn{(-\infty ,1]} if
\code{observed} and \code{predicted} are numeric vectors without NA values
and the same length. Otherwise the output will be \code{NA}.
}
\description{
\code{r2} returns a measure of how well future samples are likely to be
predicted by the model. Best possible score is 1.0 and it can be negative
(because the model can be arbitrarily worse). A constant model that always
predicts the expected value of y, disregarding the input features, would get
a score of 0.0.
}
\details{
\deqn{R^{2} = 1 - \frac{RSS}{TSS}}
where:
\deqn{RSS = \sum_{i=1}^n{y_i - \hat {y_i}}}
\deqn{TSS = \sum _{i=1}^n {(y_i - \bar y)^2}}
and \eqn{\bar y} is the overal mean and \eqn{\hat {y_i}} is the predicted
value of \eqn{y_i}.
}
\examples{
y_obs  <- c(3.0, -0.5, 2.0, 5.0)
y_pred <- c(3.9, -0.2, 2.8, 4.5)
r2(y_obs, y_pred)
}
