% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmetrics.R
\docType{package}
\name{regmetrics}
\alias{regmetrics}
\alias{regmetrics-package}
\title{regmetrics: A package for computating regression error metrics
and plots}
\description{
The \code{regmetrics} package provides three categories of important functions:
error metrics, plots and comparisons
}
\section{Error metrics}{

The error metircs functions are: Coefficient of determination,
Root-mean-square error, Mean absolute error, Explained variance score,
Mean squared logarithmic error and Mean bias error.
}

