% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{rmse}
\alias{rmse}
\title{Root-mean-square error}
\usage{
rmse(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
Is a frequently used measure of the differences between values (sample and
 population values) predicted by a model or an estimator and the values
 actually observed.
}
