% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{rsquared}
\alias{rsquared}
\title{Coefficient of determination}
\usage{
rsquared(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
It provides a measure of how well future samples are likely to be predicted
by the model. Best possible score is 1.0 and it can be negative (because the
model can be arbitrarily worse). A constant model that always predicts the
expected value of y, disregarding the input features, would get a R^2
score of 0.0.
}
