% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{tss}
\alias{tss}
\title{Compute the total sum of squares}
\usage{
tss(observed)
}
\arguments{
\item{observed}{observed  numeric vector.}
}
\value{
A numeric vector of length one.
}
\description{
\code{tss} returns the sum, over all observations, of the squared differences
of each observation from the overall mean.
}
\details{
\deqn{TSS = \sum _{i=1}^n {(y_i - \bar y)^2}}
where \eqn{ \bar{y} } is the overal mean
}
\examples{
 y_obs <- c(3, -0.5, 2, 5)
 tss(y_obs)
}
