% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{mbe}
\alias{mbe}
\title{Mean bias error}
\usage{
mbe(observed, predicted)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}
}
\value{
A numeric vector of length one. A number in
\eqn{(-\infty), +\infty)} if \code{observed} and  \code{predicted} are
numeric vectors without NA values with the same length. Otherwise the output
will be \code{NA}.
}
\description{
\code{mde} returns the direction of the error bias. Its value is related
to the magnitude of values under investigation. A negative MBE occurs when
predictions are smaller in value than observations.
}
\details{
\deqn{MBE = \frac{1}{n} \sum_{i=1}^n(y_i- \hat{y_i})}
where  \eqn{\hat {y_i}} is the predicted value of \eqn{y_i}.
}
\examples{
 y_obs  <- c(0.22,  0.83, -0.12, 0.89, -0.23, -1.30, -0.15, -1.4,
             0.62,  0.99, -0.18, 0.32,  0.34, -0.30,  0.04, -0.87,
             0.55, -1.30, -1.15, 0.20)
 y_pred <- c(0.24,  0.78, -0.66,  0.53, 0.70, -0.75, -0.41, -0.43,
               0.49,  0.79, -1.19,  0.06, 0.75, -0.07,  0.43, -0.42,
              -0.25, -0.64, -1.26, -0.07)
 mbe(y_obs, y_pred)
}
