% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{rss}
\alias{rss}
\title{Compute the residual sum of squares}
\usage{
rss(observed, predicted)
}
\arguments{
\item{observed}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
It is defined as the sum of the squares of residuals, deviations predicted
 from actual empirical values of data. It is a measure of the discrepancy
 between the data and an estimation model
}
