% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{rss}
\alias{rss}
\title{Compute the residual sum of squares}
\usage{
rss(observed, predicted)
}
\arguments{
\item{observed}{observed  numeric vector.}

\item{predicted}{predicted numeric vector.}
}
\value{
A numeric vector of length one.
}
\description{
\code{rss} returns the sum of the squares of residuals, deviations predicted
 from actual empirical values of data.
}
\details{
\deqn{RSS = \sum_{i=1}^n{y_i - \hat {y_i}}}
where \eqn{\hat {y_i}} is the predicted value of \eqn{y_i}.
}
\examples{
 y_obs  <- c(3.0, -0.5, 2.0, 5.0)
 y_pred <- c(3.1, -0.4, 1.8, 4.5)
 rss(y_obs, y_pred)
}
