% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{tss}
\alias{tss}
\title{Compute the total sum of squares}
\usage{
tss(observed)
}
\arguments{
\item{observed}{ground truth numeric vector}
}
\value{
If all observations are numeric, then the output will be a number.
Otherwise the output will be an integer
}
\description{
\code{tss} returns the sum, over all observations, of the squared differences
of each observation from the overall mean
\url{https://en.wikipedia.org/wiki/Total_sum_of_squares}.
}
\details{
\deqn{TSS = \sum _{i=1}^n {(y_i - \bar y)^2}}
}
